+++ src/components/NewNote.vue
@@ -32,7 +32,7 @@
 
 <script lang="ts" setup>
 import { reactive, ref, watch, onMounted, computed } from 'vue';
-import { CheckOutlined, TagsOutlined, PlusOutlined } from '@ant-design/icons-vue';
+import { CheckOutlined, TagsOutlined } from '@ant-design/icons-vue';
 import { useNotesStore } from '@/stores/notesStore';
 import { useTagsStore } from '@/stores/tagsStore';
 import type { NoteType } from '@/types/Note';
@@ -59,11 +59,11 @@ interface FormState {
        title: string;
        description: string;
 }
-watch(selectedTag, (newTagId) => {
-       if (newTagId) {
-              const tagData = tagsStore.tags.find(tag => tag.id === newTagId);
-       }
-});
+// watch(selectedTag, (newTagId) => {
+//        if (newTagId) {
+//               const tagData = tagsStore.tags.find(tag => tag.id === newTagId);
+//        }
+// });
 watch(formState, () => {
        noteModified.value = formState.title.trim() !== '' || formState.description.trim() !== '';
 }, { deep: true });
