+++ src/layouts/NotConnectedLayout.vue
@@ -10,6 +10,6 @@
 </template>
 
 <script setup lang="ts">
-import Header from '@/components/Header.vue';
+import Header from '@/components/AppHeader.vue';
 // Composition API setup logic, if needed
 </script>
